Add files via upload
Added dialogue for random things, but mostly regular female NPCs, especially sisters, cats and catlike monsters, most of the hidden characters (at least one is also missing from this project), the worshipable god NPCs, and the Inside Gods.  The translation is probably not super accurate, but is at least way better than what you'd  get from Google translate, and I tried to implement their puns and ambiguities.

Also, check out what happens when cows kill things. Hilarious, but I didn't take huge liberties with it - that's basically what they sound is like in the Japanese, whether that's what they were going for or not.

A big, deliberate omission was NOT having regular animal cats welcome you home in English. That's too immersion-breaking in my opinion.

Unlike the main branch of this project, the Japanese in these files is not properly encoded.